refactor: splitting up of ConnectionStats, IPTag, MinimalPacket and SessionInfo
